+++ apps/bbsync/exceptions.py
@@ -1,0 +1,9 @@
+"""
+BBSync exceptions
+"""
+
+
+class SRTNotesValidationError(Exception):
+    """
+    SRT notes JSON valition error exception class
+    """