+++ apps/bbsync/query.py
@@ -1,13 +1,14 @@
 import json
+from ast import literal_eval
 from itertools import chain
 
-from collectors.bzimport.constants import ANALYSIS_TASK_PRODUCT, BZ_DT_FMT_HISTORY
-from osidb.models import Flaw, FlawImpact, PsModule, Tracker
+import jsonschema
 
-DATE_FMT = "%Y-%m-%d"
-# these two time formats are the same
-# thus spare us defining it again
-DATETIME_FMT = BZ_DT_FMT_HISTORY
+from collectors.bzimport.constants import ANALYSIS_TASK_PRODUCT
+from osidb.models import Flaw, FlawImpact, FlawMeta, PsModule, Tracker
+
+from .constants import DATE_FMT, DATETIME_FMT, SRTNOTES_SCHEMA_PATH
+from .exceptions import SRTNotesValidationError
 
 
 class SRTNotesBuilder:
@@ -33,6 +34,7 @@ class SRTNotesBuilder:
         """
         if self._json is None:
             self.generate()
+            self.validate()
 
         return json.dumps(self._json)
 
@@ -63,33 +65,69 @@ class SRTNotesBuilder:
         generate json content
         """
         self.restore_original()
-        self.generate_impact()
-        self.generate_jira_trackers()
+        self.generate_acknowledgments()
+        self.generate_affects()
         self.generate_date("unembargo_dt", "public")
         self.generate_date("reported_dt", "reported")
+        self.generate_impact()
+        self.generate_jira_trackers()
+        # TODO the references are not yet fully implemented in OSIDB
+        # this requirement is trackerd in OSIDB-71 and when fulfilled
+        # we need to implement generate_references accordingly
         self.generate_source()
+        self.generate_string("cvss2", "cvss2")
+        self.generate_string("cvss3", "cvss3")
+        # TODO the CVSS comments are not yet supported in OSIDB and tracked as part of OSIDB-377
+        # we need to implement their SRT notes generator afterwards and probably modify
+        # the existing CVSS into SRT notes generation too
+        self.generate_string("cwe_id", "cwe")
+        # TODO mitigation field is not yet supported in OSIDB
+        # this requirement is tracked in when OSIDB-584 and when fulfilled
+        # we need to implement its proper SRT notes generator too while
+        # it should probably be enough to just uncomment the next line
+        # self.generate_string("mitigation", "mitigation")
+        self.generate_string("statement", "statement")
 
-    def generate_impact(self):
+    def generate_acknowledgments(self):
         """
-        generate impact attribute
-        """
-        impact = "none" if not self.flaw.impact else self.flaw.impact.lower()
-        self.add_conditionally("impact", impact, empty_value="none")
-
-    def generate_jira_trackers(self):
-        """
-        generate array of Jira tracker identifier pairs
-        consisting of BTS name being Jira instance identifier
-        and Jira issue key in given Jira instance
+        generate array of acknowledgments to SRT notes
         """
         self.add_conditionally(
-            "jira_trackers",
+            "acknowledgments",
             [
-                # BTS name is always jboss which is the
-                # historical naming of the only Jira instance we use
-                {"bts_name": "jboss", "key": tracker.external_system_id}
+                {
+                    "affiliation": meta.meta_attr["affiliation"],
+                    # hstore holds bolean values as strings containing True|False
+                    # so we need to explicitly convert it to the bolean value
+                    "from_upstream": literal_eval(meta.meta_attr["from_upstream"]),
+                    "name": meta.meta_attr["name"],
+                }
+                for meta in self.flaw.meta.all()
+                if meta.type == FlawMeta.FlawMetaType.ACKNOWLEDGMENT
+            ],
+        )
+
+    def generate_affects(self):
+        """
+        generate array of affects to SRT notes
+        """
+        self.add_conditionally(
+            "affects",
+            [
+                {
+                    "ps_module": affect.ps_module,
+                    "ps_component": affect.ps_component,
+                    "affectedness": affect.affectedness.lower() or None,
+                    "resolution": affect.resolution.lower() or None,
+                    # there is an interesting fact that the impact may be null or none
+                    # while these two values hold the same information (empty impact)
+                    # so let us prefer null which may cause some unexpected rewrites
+                    # from none to null but this can be considered as data fixes
+                    "impact": affect.impact.lower() or None,
+                    "cvss2": affect.cvss2 or None,
+                    "cvss3": affect.cvss3 or None,
+                }
                 for affect in self.flaw.affects.all()
-                for tracker in affect.trackers.filter(type=Tracker.TrackerType.JIRA)
             ],
         )
 
@@ -119,6 +157,30 @@ class SRTNotesBuilder:
         else:
             self._json[srtnotes_attribute] = date_value.strftime(DATETIME_FMT)
 
+    def generate_impact(self):
+        """
+        generate impact attribute
+        """
+        impact = "none" if not self.flaw.impact else self.flaw.impact.lower()
+        self.add_conditionally("impact", impact, empty_value="none")
+
+    def generate_jira_trackers(self):
+        """
+        generate array of Jira tracker identifier pairs
+        consisting of BTS name being Jira instance identifier
+        and Jira issue key in given Jira instance
+        """
+        self.add_conditionally(
+            "jira_trackers",
+            [
+                # BTS name is always jboss which is the
+                # historical naming of the only Jira instance we use
+                {"bts_name": "jboss", "key": tracker.external_system_id}
+                for affect in self.flaw.affects.all()
+                for tracker in affect.trackers.filter(type=Tracker.TrackerType.JIRA)
+            ],
+        )
+
     def generate_source(self):
         """
         generate source attribute
@@ -128,6 +190,15 @@ class SRTNotesBuilder:
         source = source if source else None
         self.add_conditionally("source", source)
 
+    def generate_string(self, flaw_attribute, srtnotes_attribute):
+        """
+        generate given string attribute
+        generic generator for string attributes with no special handling
+        """
+        self.add_conditionally(
+            srtnotes_attribute, getattr(self.flaw, flaw_attribute) or None
+        )
+
     def restore_original(self):
         """
         restore the original SRT notes attributes
@@ -136,6 +207,19 @@ class SRTNotesBuilder:
         srtnotes = self.flaw.meta_attr.get("original_srtnotes")
         self._json = json.loads(srtnotes) if srtnotes else {}
         self._original_keys = list(self._json.keys())
+
+    def validate(self):
+        """
+        validation safeguard to ensure that we always create valid SRT notes JSON data
+        throws SRTNotesValidationError exception in the case of invalid SRT notes JSON
+        """
+        with open(SRTNOTES_SCHEMA_PATH) as schema_fp:
+            self.srtnotes_schema = json.load(schema_fp)
+
+        try:
+            jsonschema.validate(self._json, schema=self.srtnotes_schema)
+        except jsonschema.ValidationError:
+            raise SRTNotesValidationError("Invalid JSON produced for SRT notes")
 
 
 class BugzillaQueryBuilder: